+++ src/share.js
@@ -1,4 +1,5 @@
 export const definitions = {};
+export const optionsCollection = {};
 
 export const stacks = [];
 
@@ -6,5 +7,5 @@ export default {
   initialized: !1,
   id: 1,
   root: document.body,
-  pushHash: !1,
+  pushingHash: !1,
 };