+++ CHANGELOG.md
@@ -1,5 +1,12 @@
 # Changelog
 
+## 1.0.1 [2019-08-05]
+
+- Can run a non route page when application is not initialized.
+- `startPage` add `options` argument.
+- Add `getRoutePage, getRoutePagesLength, getCurrentRoutePage, getPages, getRoutePages` API.
+- Remove `getLastRoutePage` API.
+
 ## 1.0.0 [2019-07-30]
 
 - Rewrite core code.